extract drag logic to hook
